Fix payload type and range

+++ Neurophotometrics.Design/LedCalibrationEditor.cs
@@ -50,7 +50,7 @@ namespace Neurophotometrics.Design
             return Observable.FromEventPattern(
                 handler => slider.ValueChanged += handler,
                 handler => slider.ValueChanged -= handler)
-                .Select(evt => HarpCommand.WriteByte(address, (byte)slider.Value));
+                .Select(evt => HarpCommand.WriteUInt16(address, (ushort)slider.Value));
         }
 
         public IObservable<HarpMessage> Commands { get; private set; }
